+++ mne_pipeline_hd/gui/gui_utils.py
@@ -9,16 +9,19 @@ License: BSD (3-clause)
 """
 import io
 import logging
-import smtplib
-import ssl
 import sys
 import traceback
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
 
-from PyQt5.QtCore import QObject, QRunnable, Qt, pyqtSignal, pyqtSlot
-from PyQt5.QtWidgets import QApplication, QDesktopWidget, QDialog, QGridLayout, QLabel, QLineEdit, QMessageBox, \
-    QPushButton
+from PyQt5.QtCore import QObject, QRunnable, pyqtSignal, pyqtSlot
+from PyQt5.QtWidgets import QApplication
+
+
+def get_ratio_geometry(size_ratio):
+    desk_geometry = QApplication.instance().desktop().availableGeometry()
+    height = int(desk_geometry.height() * size_ratio)
+    width = int(desk_geometry.width() * size_ratio)
+
+    return width, height
 
 
 def get_exception_tuple():
@@ -119,110 +122,3 @@ class StderrStream(io.TextIOBase):
                 self.signal.text_written.emit(text)
                 self.last_text = text
 
-
-class ErrorDialog(QDialog):
-    def __init__(self, exception_tuple, parent=None, title=None):
-        if parent:
-            super().__init__(parent)
-        else:
-            super().__init__()
-        self.err = exception_tuple
-        self.title = title
-        if self.title:
-            self.setWindowTitle(self.title)
-        else:
-            self.setWindowTitle('An Error ocurred!')
-
-        self.init_ui()
-
-        if parent:
-            self.open()
-        else:
-            self.show()
-        self.center()
-        self.raise_win()
-
-    def init_ui(self):
-        layout = QGridLayout()
-
-        self.label = QLabel()
-        self.formated_tb_text = self.err[2].replace('\n', '<br>')
-        if self.title:
-            self.html_text = f'<h1>{self.title}</h1>' \
-                             f'<h2>{self.err[1]}</h2>' \
-                             f'{self.formated_tb_text}'
-        else:
-            self.html_text = f'<h1>{self.err[1]}</h1>' \
-                             f'{self.formated_tb_text}'
-        self.label.setText(self.html_text)
-        layout.addWidget(self.label, 0, 0, 1, 2)
-
-        self.name_le = QLineEdit()
-        self.name_le.setPlaceholderText('Enter your Name (optional)')
-        layout.addWidget(self.name_le, 1, 0)
-
-        self.email_le = QLineEdit()
-        self.email_le.setPlaceholderText('Enter your E-Mail-Adress (optional)')
-        layout.addWidget(self.email_le, 1, 1)
-
-        self.send_bt = QPushButton('Send Error-Report')
-        self.send_bt.clicked.connect(self.send_report)
-        layout.addWidget(self.send_bt, 2, 0)
-
-        self.close_bt = QPushButton('Close')
-        self.close_bt.clicked.connect(self.close)
-        layout.addWidget(self.close_bt, 2, 1)
-
-        self.setLayout(layout)
-
-        self.desk_geometry = QApplication.instance().desktop().availableGeometry()
-        self.size_ratio = 0.7
-        height = int(self.desk_geometry.height() * self.size_ratio)
-        width = int(self.desk_geometry.width() * self.size_ratio)
-        self.setMaximumSize(width, height)
-
-    def center(self):
-        qr = self.frameGeometry()
-        cp = QDesktopWidget().availableGeometry().center()
-        qr.moveCenter(cp)
-        self.move(qr.topLeft())
-
-    def raise_win(self):
-        if sys.platform == 'win32':
-            # on windows we can raise the window by minimizing and restoring
-            self.showMinimized()
-            self.setWindowState(Qt.WindowActive)
-            self.showNormal()
-        else:
-            # on osx we can raise the window. on unity the icon in the tray will just flash.
-            self.activateWindow()
-            self.raise_()
-
-    def send_report(self):
-        msg_box = QMessageBox.question(self, 'Send an E-Mail-Bug-Report?',
-                                       'Do you really want to send an E-Mail-Report?',
-                                       QMessageBox.Yes | QMessageBox.No, QMessageBox.No)
-        if msg_box == QMessageBox.Yes:
-            port = 465
-            adress = 'mne.pipeline@gmail.com'
-            password = '24DecodetheBrain7!'
-
-            context = ssl.create_default_context()
-
-            message = MIMEMultipart("alternative")
-            message['Subject'] = str(self.err[1])
-            message['From'] = adress
-            message["To"] = adress
-
-            message_body = MIMEText(f'<b><big>{self.name_le.text()}</b></big><br>'
-                                    f'<i>{self.email_le.text()}</i><br><br>'
-                                    f'<b>{sys.platform}</b><br>{self.formated_tb_text}', 'html')
-            message.attach(message_body)
-            try:
-                with smtplib.SMTP_SSL('smtp.gmail.com', port, context=context) as server:
-                    server.login('mne.pipeline@gmail.com', password)
-                    server.sendmail(adress, adress, message.as_string())
-                QMessageBox.information(self, 'E-Mail sent', 'An E-Mail was sent to mne.pipeline@gmail.com\n'
-                                                             'Thank you for the Report!')
-            except OSError:
-                QMessageBox.information(self, 'E-Mail not sent', 'Sending an E-Mail is not possible on your OS')
